Fix bug in script.js
When the player stops a game in the middle of a clue sequence and starts a new game, the next button in the clue sequence from the previous game lights up at the start of the new game. 

Set buttonPlaying to null when button is cleared. Set 'buttonPlaying', 'timer', 'startClock', 'playClue', and 'enable' to null in startGame to prevent any values from previous game from carrying over to new game.